# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CryptoVault Exchange & Shop API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/users:
    get:
      operationId: adminListUsers
      tags: [admin]
      summary: List all users (admin)
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 50 }
        - name: offset
          in: query
          schema: { type: integer, default: 0 }
        - name: search
          in: query
          schema: { type: string }
      responses:
        "200":
          description: User list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminUser"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
  /admin/users/{userId}/block:
    patch:
      operationId: adminToggleUserBlock
      tags: [admin]
      summary: Toggle withdrawal block on user
      parameters:
        - name: userId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlockToggle"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
  /admin/users/{userId}/add-balance:
    post:
      operationId: adminAddUserBalance
      tags: [admin]
      summary: Credit wallet balance for a user (admin)
      parameters:
        - name: userId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddBalanceRequest"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: User not found
  /admin/deposits:
    get:
      operationId: adminListDeposits
      tags: [admin]
      summary: List all deposits (admin)
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 50 }
        - name: offset
          in: query
          schema: { type: integer, default: 0 }
        - name: status
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Deposit list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Deposit"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
  /admin/withdrawals:
    get:
      operationId: adminListWithdrawals
      tags: [admin]
      summary: List all withdrawals (admin)
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 50 }
        - name: offset
          in: query
          schema: { type: integer, default: 0 }
        - name: status
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Withdrawal list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Withdrawal"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
  /admin/withdrawals/{id}/approve:
    post:
      operationId: adminApproveWithdrawal
      tags: [admin]
      summary: Approve a pending withdrawal
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Withdrawal approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Withdrawal"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
  /admin/withdrawals/{id}/reject:
    post:
      operationId: adminRejectWithdrawal
      tags: [admin]
      summary: Reject a pending withdrawal
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Withdrawal rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Withdrawal"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
...
  /admin/settings:
    get:
      operationId: adminGetSettings
      tags: [admin]
      summary: Get all platform settings (admin)
      responses:
        "200":
          description: All settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlatformSettings"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
    put:
      operationId: adminUpdateSettings
      tags: [admin]
      summary: Update platform settings (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SettingsUpdate"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlatformSettings"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
  /admin/stats:
    get:
      operationId: adminGetStats
      tags: [admin]
      summary: Get platform stats (admin)
      responses:
        "200":
          description: Platform stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden

  # ─── NFT / V2 Token Economy ──────────────────────────────────────────────────
...
components:
  schemas:
...
    AdminUser:
      type: object
      required: [id, clerkId, email, walletBalance, earningsBalance, depositAddress, referralCode, isAdmin, withdrawalBlocked, createdAt]
      properties:
        id: { type: string }
        clerkId: { type: string }
        email: { type: string }
        fullName: { type: ["string", "null"] }
        walletBalance: { type: string }
        earningsBalance: { type: string }
        depositAddress: { type: string }
        referralCode: { type: string }
        isAdmin: { type: boolean }
        withdrawalBlocked: { type: boolean }
        createdAt: { type: string, format: date-time }
        totalDeposited: { type: string }
        totalWithdrawn: { type: string }
...
    Deposit:
      type: object
      required: [id, userId, amount, fee, netAmount, status, txHash, createdAt]
      properties:
        id: { type: string }
        userId: { type: string }
        amount: { type: string }
        fee: { type: string }
        netAmount: { type: string }
        status:
          type: string
          enum: [pending, sweeping, credited, failed]
        txHash: { type: ["string", "null"] }
        sweepTxHash: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
        creditedAt: { type: ["string", "null"] }
...
    Withdrawal:
      type: object
      required: [id, userId, amount, fee, netAmount, destinationAddress, status, createdAt]
      properties:
        id: { type: string }
        userId: { type: string }
        amount: { type: string }
        fee: { type: string }
        netAmount: { type: string }
        destinationAddress: { type: string }
        status:
          type: string
          enum: [pending, approved, rejected]
        txHash: { type: ["string", "null"] }
        rejectionReason: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
        processedAt: { type: ["string", "null"] }
...
    PlatformSettings:
      type: object
      required: [minDepositUsdt, depositFeeFlat, depositFeePercent, withdrawFeeFlat, withdrawFeePercent, withdrawFeeMode, withdrawalMode, withdrawalEnabled, otpWithdrawalEnabled, adminMasterWallet, bscRpcUrl]
      properties:
        minDepositUsdt: { type: string }
        depositFeeFlat: { type: string }
        depositFeePercent: { type: string }
        withdrawFeeFlat: { type: string }
        withdrawFeePercent: { type: string }
        withdrawFeeMode:
          type: string
          enum: [deduct_from_amount, deduct_from_balance]
        withdrawalMode:
          type: string
          enum: [manual, auto]
        withdrawalEnabled: { type: boolean }
        otpWithdrawalEnabled: { type: boolean }
        adminMasterWallet: { type: string }
        bscRpcUrl: { type: string }
        gasWalletAddress: { type: ["string", "null"] }
        telegramBotToken: { type: string }
        telegramBotUsername: { type: string }
    SettingsUpdate:
      type: object
      properties:
        minDepositUsdt: { type: string }
        depositFeeFlat: { type: string }
        depositFeePercent: { type: string }
        withdrawFeeFlat: { type: string }
        withdrawFeePercent: { type: string }
        withdrawFeeMode:
          type: string
          enum: [deduct_from_amount, deduct_from_balance]
        withdrawalMode:
          type: string
          enum: [manual, auto]
        withdrawalEnabled: { type: boolean }
        otpWithdrawalEnabled: { type: boolean }
        adminMasterWallet: { type: string }
        bscRpcUrl: { type: string }
        gasWalletPrivateKey: { type: string }
        withdrawWalletPrivateKey: { type: string }
        telegramBotToken: { type: string }
        telegramBotUsername: { type: string }
...
    BlockToggle:
      type: object
      required: [blocked]
      properties:
        blocked: { type: boolean }
    AddBalanceRequest:
      type: object
      required: [amount]
      properties:
        amount:
          type: string
          description: USDT amount to credit (must be positive)
        note:
          type: string
          description: Optional admin note
    AdminStats:
      type: object
      required: [totalUsers, totalDeposited, totalWithdrawn, pendingWithdrawals, totalP2P, depositCount, withdrawalCount]
      properties:
        totalUsers: { type: integer }
        totalDeposited: { type: string }
        totalWithdrawn: { type: string }
        pendingWithdrawals: { type: string }
        totalP2P: { type: string }
        depositCount: { type: integer }
        withdrawalCount: { type: integer }
        recentDeposits:
          type: array
          items: { $ref: "#/components/schemas/Deposit" }
        recentWithdrawals:
          type: array
          items: { $ref: "#/components/schemas/Withdrawal" }

    # ─── Shop schemas ─────────────────────────────────────────────────────────
...